test: isolate security header coverage
