+++ apps/api/src/routes/__tests__/cors.test.ts
@@ -72,16 +72,6 @@ describe("CORS configuration", () => {
         };
     });
 
-    it("sets frame blocking headers consistently", async () => {
-        const app = await createTestApp();
-        const env = createTestEnv();
-
-        const res = await app.request("http://localhost/", {}, env as any);
-
-        expect(res.headers.get("x-frame-options")).toBe("DENY");
-        expect(res.headers.get("content-security-policy")).toContain("frame-ancestors 'none'");
-    });
-
     it("uses FRONTEND_URL as fallback when ALLOWED_ORIGINS is not set", async () => {
         const app = await createTestApp();
         const env = createTestEnv({
